DevFlow: add EnableDevFlow/DEVFLOW + .devflow config to Agent targets
- New EnableDevFlow MSBuild property (default true in Debug, false
  otherwise) gates the dedicated DEVFLOW compiler symbol that is
  appended to DefineConstants by the Agent package's own .targets.
- DevFlowConstant property lets users override the symbol name.
- .targets now resolves the port from -p:DevFlowPort (or legacy
  -p:MauiDevFlowPort), then a .devflow JSON file, falling back to
  .mauidevflow with a one-time MSBuild warning asking for a rename.
- Parallel assembly metadata: emits both DevFlow* and legacy
  Microsoft.Maui.DevFlow* entries plus a new DevFlowEnabled marker.
- Gtk Agent package now packs build/buildTransitive targets so GTK
  apps get the same DEVFLOW / EnableDevFlow / .devflow treatment.
- Tests renamed to DevFlowAgentTargetsTests with new coverage for
  EnableDevFlow, DevFlowConstant, config-file precedence, legacy
  fallback warning, and legacy MauiDevFlowPort compatibility.
